feat: move buttons files to nested directory Inputs
