--- src/components/Button/button.scss
@@ -1,36 +1,0 @@
-@import '../../styles/mixins/button/button.scss';
-@import '../../styles/mixins/button/button-sizes.scss';
-@import '../../styles/mixins/button/button-variants.scss';
-
-.Button {
-  @include button-variants;
-  @include button-size;
-  @include button;
-  border-radius: 4px;
-
-  &__full-width {
-    width: 100%;
-  }
-
-  &__icon {
-    &--small {
-      height: 12px;
-    }
-
-    &--medium {
-      height: 14px;
-    }
-
-    &--large {
-      height: 16px;
-    }
-
-    &--start {
-      margin-right: 10px;
-    }
-
-    &--end {
-      margin-left: 10px;
-    }
-  }
-}
