--- src/components/ButtonGroup/stories/BasicButtonGroup.stories.tsx
@@ -1,75 +1,0 @@
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-
-// components
-import Button from '../../Button/Button';
-import ButtonGroup from '../ButtonGroup';
-import StoryComponent, {
-  ContentGridFlow,
-} from '../../../stories/components/StoryComponent/StoryComponent';
-
-// others
-import { basicButtonGroup } from '../../../stories/constants';
-import { libraryName } from '../../../constants';
-import { TStoryBlockCode } from '../../../stories/components/StoryBlockCode/types';
-
-const description = [
-  'The ButtonGroup component can be used to group related buttons.',
-];
-
-const blockCodeData: TStoryBlockCode = {
-  componentName: 'ButtonGroup',
-  props: [
-    {
-      children: Array.from(Array(3), () => '&lt;Button&gtOne&lt;/Button&gt'),
-    },
-  ],
-  imports: [
-    {
-      items: '{ Button, ButtonGroup }',
-      path: libraryName,
-    },
-  ],
-};
-
-export default {
-  component: ButtonGroup,
-  parameters: {
-    design: {
-      type: 'figma',
-      url: 'https://www.figma.com/file/yhTZ31Wn16kIrEeoHHV176/Components---UI?node-id=232%3A2',
-    },
-  },
-  title: basicButtonGroup,
-} as ComponentMeta<typeof ButtonGroup>;
-
-const Template: ComponentStory<typeof ButtonGroup> = (args) => (
-  <StoryComponent
-    blockCodeData={blockCodeData}
-    contentGridFlow={ContentGridFlow.row}
-    description={description}
-    title="Basic button group"
-  >
-    <ButtonGroup {...args}>
-      <Button>One</Button>
-      <Button>Two</Button>
-      <Button>Three</Button>
-    </ButtonGroup>
-    <ButtonGroup forcedHover {...args}>
-      <Button>One</Button>
-      <Button>Two</Button>
-      <Button>Three</Button>
-    </ButtonGroup>
-
-    <ButtonGroup disabled {...args}>
-      <Button>One</Button>
-      <Button>Two</Button>
-      <Button>Three</Button>
-    </ButtonGroup>
-  </StoryComponent>
-);
-
-export const BasicButtonGroup = Template.bind({});
-
-BasicButtonGroup.argTypes = {};
-
-BasicButtonGroup.args = {};
