+++ src/components/Inputs/index.ts
@@ -1,0 +1,3 @@
+export { Button } from './Button/Button';
+export { ButtonGroup } from './ButtonGroup/ButtonGroup';
+export { ButtonIcon } from './ButtonIcon/ButtonIcon';
