+++ src/components/index.ts
@@ -1,2 +1,1 @@
-export { Button } from './Button/Button';
-export { ButtonIcon } from './ButtonIcon/ButtonIcon';
+export * from './Inputs';
